Misc fixes
